+++ packages/init/src/webframeworks/next.ts
@@ -22,6 +22,7 @@ const nextDescription: WebFrameworkDescription = {
     federationFile: "federation/index.ts",
     loggingFile: "logging.ts",
     files: {
+      "instrumentation.ts": await readTemplate("next/instrumentation.ts"),
       "middleware.ts": await readTemplate("next/middleware.ts"),
       ...(pm !== "deno" && {
         "eslint.config.ts": await readTemplate("defaults/eslint.config.ts"),